list, detail에 작성자 이름, include 사용법

+++ src/main/webapp/jsp/article/detail.jsp
@@ -6,6 +6,7 @@
     
 <%
 	Map<String, Object> articleMap = (Map<String, Object>) request.getAttribute("articleMap");
+	int loginedMemberId = (int) request.getAttribute("loginedMemberId");
 %>
 <!DOCTYPE html>
 <html>
@@ -19,13 +20,16 @@
 	<div>번호 : <%= (int) articleMap.get("id") %></div>
 	<div>작성일 : <%= (LocalDateTime) articleMap.get("regDate") %></div>
 	<div>수정일 : <%= (LocalDateTime) articleMap.get("updateDate") %></div>
+	<div>작성자 : <%= (String) articleMap.get("writerName") %></div>
 	<div>제목 : <%= (String) articleMap.get("title") %></div>
 	<div>내용 : <%= (String) articleMap.get("body") %></div>
 	
 	<div>
 		<a href="list">목록</a>
-		<a href="modify?id=<%= (int) articleMap.get("id") %>">수정</a>
-		<a href="delete?id=<%= (int) articleMap.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		<% if (loginedMemberId != -1 && loginedMemberId == (int) articleMap.get("memberId")) { %>
+			<a href="modify?id=<%= (int) articleMap.get("id") %>">수정</a>
+			<a href="delete?id=<%= (int) articleMap.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		<% } %>
 	</div>
 </body>
 </html>
